MMP-0006 - convert AuthContext to TS
