+++ src/context/AuthContext.tsx
@@ -1,0 +1,59 @@
+import React, { createContext, useContext, useState, ReactNode } from 'react';
+
+interface User {
+  name: string;
+  email: string;
+  password: string;
+}
+
+interface LoginData {
+  email: string;
+  password: string;
+}
+
+interface AuthContextType {
+  user: User | null;
+  login: (loginData: LoginData) => boolean;
+  signup: (userData: User) => void;
+  logout: () => void;
+}
+
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+let storedUsers: Record<string, User> = {};
+
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
+  const [user, setUser] = useState<User | null>(null);
+
+  const login = (loginData: LoginData): boolean => {
+    const userData = storedUsers[loginData.email];
+    if (userData && userData.password === loginData.password) {
+      setUser(userData);
+      return true;
+    }
+    return false;
+  };
+
+  const signup = (userData: User): void => {
+    storedUsers[userData.email] = userData;
+    setUser(userData);
+  };
+
+  const logout = (): void => {
+    setUser(null);
+  };
+
+  return (
+    <AuthContext.Provider value={{ user, login, signup, logout }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
+
+export const useAuth = (): AuthContextType => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
